cell_analysis/requirements.gpu.txt
pytorch-lightning==2.0.0
timm==0.9.2
numpy==1.26.4
opencv-python-headless
albumentations==1.3.1
tifffile==2023.7.10
scikit-learn==1.3.0
tensorboard
pyyaml==6.0.1
